Refactor email sending as shared helper
